refactor: action hook 분리
